move db test utils to internal
