+++ beacon-chain/attestation/BUILD.bazel
@@ -18,7 +18,7 @@ go_test(
     srcs = ["service_test.go"],
     embed = [":go_default_library"],
     deps = [
-        "//beacon-chain/testutil:go_default_library",
+        "//beacon-chain/internal:go_default_library",
         "//beacon-chain/types:go_default_library",
         "//shared/testutil:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
