+++ beacon-chain/chaintest/BUILD.bazel
@@ -2,12 +2,14 @@ load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "format.go",
-        "main.go",
-    ],
+    srcs = ["main.go"],
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/chaintest",
     visibility = ["//visibility:private"],
+    deps = [
+        "//beacon-chain/chaintest/backend:go_default_library",
+        "@com_github_go_yaml_yaml//:go_default_library",
+        "@com_github_sirupsen_logrus//:go_default_library",
+    ],
 )
 
 go_binary(
