+++ beacon-chain/internal/BUILD.bazel
@@ -6,11 +6,13 @@ go_library(
     name = "go_default_library",
     srcs = [
         "beacon_service_mock.go",
+        "db_test_util.go",
         "validator_service_mock.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/internal",
     visibility = ["//beacon-chain:__subpackages__"],
     deps = [
+        "//beacon-chain/db:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//proto/beacon/rpc/v1:go_default_library",
         "@com_github_golang_mock//gomock:go_default_library",
