+++ beacon-chain/internal/db_test_util.go
@@ -1,4 +1,4 @@
-package testutil
+package internal
 
 import (
 	"os"
