+++ beacon-chain/sync/BUILD.bazel
@@ -23,7 +23,7 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//beacon-chain/db:go_default_library",
-        "//beacon-chain/testutil:go_default_library",
+        "//beacon-chain/internal:go_default_library",
         "//beacon-chain/types:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//shared/event:go_default_library",
